Calculate state and federal taxes using TAXSIM 32.

This function calculates state and federal income taxes using the TAXSIM 32 tax simulator. See https://users.nber.org/~taxsim/taxsim32/ for more information on TAXSIM 32.

Usage

taxsim_calculate_taxes(
  .data,
  return_all_information = FALSE,
  upload_method = "ftp"
)

Arguments

.data

Data frame containing the information that will be used to calculate taxes. This data set will be sent to TAXSIM. Data frame must have specified column names and data types.

return_all_information

Boolean (TRUE or FALSE). Whether to return all information from TAXSIM (TRUE), or only key information (FALSE). Returning all information returns 42 columns of output, while only returning key information returns 9 columns. It is faster to download results with only key information.

upload_method

Either 'ftp' or 'ssh', can also use upper case. Defaults to 'ftp'. Determines whether ftp or ssh will be used to send data to TAXSIM and retrieve the results. SSH is faster, so use it when there are over 100,000 records. SSH is available in Windows 10 since autumn of 2019.

Value

Returns a data frame with the following columns: id_number The unique id number for the row that corresponds to the id number in .datafederal_taxes Total federal taxes due. Corresponds to fiitax in TAXSIM. state_taxes Total state taxes due; if state taxes were calculated. Corresponds to siitax in TAXSIM. fica_taxes Total FICA taxes due. This includes both the employee and company share. Corresponds to fica in TAXSIM. federal_marginal_rate Federal marginal tax rate of taxpayer. Corresponds to frate in TAXSIM. state_marginal_rate State marginal tax rate of taxpayer. Corresponds to srate in TAXSIM. fica_rate FICA rate. Corresponds to ficar in TAXSIM.

Required columns

id_number An arbitrary, non-negative, whole number. Each number must be unique. This number links the results from TAXSIM 32 to the original data frame with entries.

tax_year Tax year ending Dec 31 (4 digits between 1960 and 2023). State must be zero if year is before 1977 or after 2023.

filing_status Filing status of tax unit. One of the following: "single" for single; "married, jointly" for married, filing jointly; "married, separately" for married, filing separately; "dependent child" for dependent, usually a child with income; or "head of household" for head of household filing status.

Optional columns

state State two letter abbreviation or state SOI code. If state income taxes are not needed, either label as "No State" or remove this variable. State income tax information is only available from 1977 to 2023.

primary_age Age of primary taxpayer as of December 31st of tax year. Taxpayer age variables determine eligibility for additional standard deductions, personal exemption, EITC and AMT exclusion.

spouse_age Age of spouse as of December 31st of tax year.

num_dependents Total number of dependents.

num_dependents_thirteen Number of children under 13 with eligible child care expenses (Dependent Care Credit).

num_dependents_seventeen Number of children under 17 for the entire tax year (Child Credit). This includes children under 13.

num_dependents_eitc Number of qualifying children for EITC. (Typically younger than 19 or younger than 24 and a full-time student).

primary_wages Wage and salary income of Primary Taxpayer (include self-employment but no QBI).

spouse_wages Wage and salary income of Spouse (include self-employment but no QBI). Must be zero or the column should not exist for non-joint returns.

dividends Dividend income (qualified dividends only for 2003 on).

interest Interest income received (+/-).

short_term_capital_gains Short Term Capital Gains or losses (+/-).

long_term_capital_gains Long Term Capital Gains or losses (+/-).

other_property_income Other property income subject to NIIT, including: unearned or limited partnership and passive S-Corp profits; rent not eligible for QBI deduction; non-qualified dividends; capital gains distributions on form 1040; and other income or loss not otherwise enumerated here.

other_non_property_income Other non-property income not subject to Medicare NIIT such as: alimony; nonwage fellowships; and state income tax refunds (itemizers only). Also includes adjustments and items such as: alimony paid; Keogh and IRA contributions; foreign income exclusion; and NOLs

pensions Taxable Pensions and IRA distributions.

social_security Gross Social Security Benefits.

unemployment Unemployment compensation received.

other_transfer_income Other non-taxable transfer income such as: welfare; workers comp; veterans benefits; and child support that would affect eligibility for state property tax rebates but would not be taxable at the federal level.

rent_paid Rent paid (used only for calculating state property tax rebates).

property_taxes Real Estate taxes paid. This is a preference for the AMT and is is also used to calculate state property tax rebates.

other_itemized_deductions Other Itemized deductions that are a preference for the Alternative Minimum Tax. These would include: Other state and local taxes (line 8 of Schedule A) plus local income tax; Preference share of medical expenses; Miscellaneous (line 27).

child_care_expenses Child care expenses.

misc_deductions Deductions not included in other_itemized_deductions and not a preference for the AMT, including (on Schedule A for 2009). Might include: Deductible medical expenses not included in Line 16; Motor Vehicle Taxes paid; Home mortgage interest; Charitable contributions; and Casulty or Theft Losses.

scorp_income Active S-Corp income (is SSTB).

qualified_business_income Primary Taxpayer's Qualified Business Income (QBI) subject to a preferential rate without phaseout and assuming sufficient wages paid or capital to be eligible for the full deduction. Subject to SECA and Medicare additional Earnings Tax.

specialized_service_trade Primary Taxpayer's Specialized Service Trade or Business service (SSTB) with a preferential rate subject to claw-back. Subject to SECA and Medicare Additional Earnings Tax.

spouse_qualified_business_income Spouse's QBI. Must be zero for non-joint returns, or the column should not exist.

spouse_specialized_service_trade Spouse's SSTB. Must be zero for non-joint returns, or the column should not exist.

Note on number of dependents

num_dependents columns are not mutually exclusive. For example, a family with a 13 year old can report the dependent in num_dependents_thirteen and also in num_dependents_seventeen.

Additional Output

NA

Giving credit where it is due

The NBER's TAXSIM 32 tax simulator does all tax calculations. This package simply lets users interact with the tax simulator through R. Therefore, users should cite the TAXSIM 32 tax simulator when they use this package in their work:

Feenberg, Daniel Richard, and Elizabeth Coutts, An Introduction to the TAXSIM Model, Journal of Policy Analysis and Management vol 12 no 1, Winter 1993, pages 189-194.

Examples

family_income <- data.frame(
    id_number = c(1, 2),
     state = c('North Carolina', 'NY'),
     tax_year = c(2015, 2015),
     filing_status = c('single', 'married, jointly'),
     primary_wages = c(10000, 100000),
     primary_age = c(26, 36)
)

family_taxes <- taxsim_calculate_taxes(family_income, upload_method = 'ftp')
#> [1] "All required columns are present and the data is in the proper format!"
#> [1] "Uploading data to TAXSIM server via ftp."
#> [1] "Downloading data from TAXSIM server via ftp."

merge(family_income, family_taxes, by = 'id_number')
#>   id_number          state tax_year    filing_status primary_wages primary_age
#> 1         1 North Carolina     2015           single         1e+04          26
#> 2         2             NY     2015 married, jointly         1e+05          36
#>   federal_taxes state_taxes fica_taxes federal_marginal_rate
#> 1       -369.12      143.75       1530                  7.65
#> 2      11437.50     4761.98      15300                 25.00
#>   state_marginal_rate fica_rate
#> 1                5.75        15
#> 2                6.45        15